reorg trak/models into separate files for clarity, flake8 ignore */__init__.py:F401 (unused import)
