+++ setup.cfg
@@ -33,6 +33,7 @@ max-complexity = 12
 ignore = E501, W503, E203
 per-file-ignores =
     apps/accounts/apps.py:F401
+    */__init__.py:F401
 
 [isort]
 skip_glob = *_alter_*.py, *_initial.py
